+++ bootique/src/main/java/io/bootique/meta/application/ApplicationMetadata.java
@@ -23,10 +23,7 @@ import io.bootique.BootiqueException;
 import io.bootique.meta.MetadataNode;
 import io.bootique.meta.config.ConfigValueMetadata;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.*;
 
 /**
  * Metadata object representing current application and its command-line interface.
@@ -35,11 +32,11 @@ public class ApplicationMetadata implements MetadataNode {
 
     private String name;
     private String description;
-    private final Collection<CommandMetadata> commands;
-    private final Collection<OptionMetadata> options;
+    private final List<CommandMetadata> commands;
+    private final List<OptionMetadata> options;
     // a combination of "commands" and "options"
-    private final Collection<OptionMetadata> cliOptions;
-    private final Collection<ConfigValueMetadata> variables;
+    private final List<OptionMetadata> cliOptions;
+    private final List<ConfigValueMetadata> variables;
 
     private ApplicationMetadata() {
         this.commands = new ArrayList<>();
@@ -80,7 +77,9 @@ public class ApplicationMetadata implements MetadataNode {
 
     /**
      * Returns a combination of commands and options as a single collection of OptionMetadata. This is a view that
-     * the users sees on the command line.
+     * the users sees on the command line. Options "shortName" property in this collection may differ from that of
+     * the original commands or options, as it is adjusted for the application context and the presence of conflicting
+     * names.
      *
      * @since 3.0
      */
@@ -106,11 +105,16 @@ public class ApplicationMetadata implements MetadataNode {
         }
 
         public ApplicationMetadata build() {
-            checkCliOptionNameDuplicates();
+            throwOnConflictingFullNames();
+            rewriteConflictingShortNames();
+
+            // set the canonical alphabetic order for display in help, etc.
+            Collections.sort(application.cliOptions, Comparator.comparing(OptionMetadata::getName));
+
             return application;
         }
 
-        private void checkCliOptionNameDuplicates() {
+        private void throwOnConflictingFullNames() {
             if (application.cliOptions.size() > 1) {
                 Set<String> distinctNames = new HashSet<>();
                 application.cliOptions.forEach(om -> {
@@ -119,6 +123,62 @@ public class ApplicationMetadata implements MetadataNode {
                     }
                 });
             }
+        }
+
+        private void rewriteConflictingShortNames() {
+            // We will only disable conflicting short names in the "cliOptions" collection , leaving short names
+            // intact in the original commands and options metadata.
+
+            int len = application.cliOptions.size();
+            Map<String, List<Integer>> shortNames = new HashMap<>();
+            for (int i = 0; i < len; i++) {
+                shortNames.computeIfAbsent(application.cliOptions.get(i).getShortName(), sn -> new ArrayList<>(3)).add(i);
+            }
+
+            // wipe out short names conflicting with other short names
+            for (Map.Entry<String, List<Integer>> e : shortNames.entrySet()) {
+
+                int slen = e.getValue().size();
+
+                // disable short options if there are multiple overlapping options
+                if (slen > 1) {
+                    for (int i = 0; i < slen; i++) {
+                        int oi = e.getValue().get(i);
+                        OptionMetadata oldOpt = application.cliOptions.get(oi);
+                        application.cliOptions.set(oi, sansShortName(oldOpt));
+                    }
+
+                    // clear wiped out short names so that the next check against full names doesn't generate a
+                    // conflict for the names already gone
+                    e.getValue().clear();
+                }
+            }
+
+            // wipe out short names conflicting with full names
+            for (OptionMetadata o : application.cliOptions) {
+
+                if (o.getName().length() == 1) {
+                    List<Integer> conflicting = shortNames.getOrDefault(o.getName(), List.of());
+
+                    // we no longer have multiple short name groups in "shortNames", so the size can only be 0 or 1
+                    if (conflicting.size() == 1) {
+                        int i = conflicting.get(0);
+                        OptionMetadata oldOpt = application.cliOptions.get(i);
+                        application.cliOptions.set(i, sansShortName(oldOpt));
+                    }
+                }
+            }
+        }
+
+        private OptionMetadata sansShortName(OptionMetadata md) {
+            return new OptionMetadata(
+                    md.getName(),
+                    md.getDescription(),
+                    null,
+                    md.getValueCardinality(),
+                    md.getValueName(),
+                    md.getDefaultValue()
+            );
         }
 
         public Builder name(String name) {